[build] #54 - Redesign directory architecture
