+++ src/modules/linkings.module.ts
@@ -1,14 +1,14 @@
 import { Module } from '@nestjs/common';
-import { LinkingsController } from './linkings.controller';
-import { LinkingsService } from './linkings.service';
-import { TypeormRepositoryModule } from '../../database/typeorm-repository.module';
+import { LinkingsController } from '../controllers/linkings.controller';
+import { LinkingsService } from '../services/linkings.service';
+import { TypeormRepositoryModule } from './typeorm-repository.module';
 import {
   LinkingImageRepository,
   LinkingLikeRepository,
   LinkingRepository,
   LinkingRequestRepository,
-} from '../../repositories/linking.repository';
-import { HashTagRepository } from '../../repositories/hash-tag.repository';
+} from '../repositories/linking.repository';
+import { HashTagRepository } from '../repositories/hash-tag.repository';
 
 @Module({
   imports: [
