+++ src/modules/users.module.ts
@@ -4,27 +4,27 @@ import {
   NestModule,
   RequestMethod,
 } from '@nestjs/common';
-import { UsersController } from './users.controller';
-import { UsersService } from './users.service';
-import { TypeormRepositoryModule } from '../../database/typeorm-repository.module';
-import { UserRepository } from '../../repositories/user.repository';
+import { UsersController } from '../controllers/users.controller';
+import { UsersService } from '../services/users.service';
+import { TypeormRepositoryModule } from './typeorm-repository.module';
+import { UserRepository } from '../repositories/user.repository';
 import {
   EstateLikeRepository,
   EstateRepository,
-} from '../../repositories/estate.repository';
+} from '../repositories/estate.repository';
 import {
   OpinionRepository,
   ProposalLikeRepository,
   ProposalRepository,
-} from '../../repositories/proposal.repository';
+} from '../repositories/proposal.repository';
 import {
   LinkingLikeRepository,
   LinkingRepository,
   LinkingRequestRepository,
-} from '../../repositories/linking.repository';
+} from '../repositories/linking.repository';
 import { JwtModule } from '@nestjs/jwt';
 import { ConfigModule, ConfigService } from '@nestjs/config';
-import { MulterUserIdMiddleware } from '../../middleware/multer-user-id.middleware';
+import { MulterUserIdMiddleware } from '../middleware/multer-user-id.middleware';
 
 @Module({
   imports: [
