+++ src/repositories/estate.repository.ts
@@ -1,4 +1,4 @@
-import { TypeormRepository } from '../database/typeorm-repository.decorator';
+import { TypeormRepository } from '../custom/decorator/typeorm-repository.decorator';
 import { Repository } from 'typeorm';
 import {
   EstateEntity,
@@ -6,7 +6,7 @@ import {
   EstateLikeEntity,
   MapNumberingEntity,
 } from '../database/entities';
-import { CreateEstateDto } from '../modules/estates/dto/req/create-estate.dto';
+import { CreateEstateDto } from '../dto/req/create-estate.dto';
 
 @TypeormRepository(EstateEntity)
 export class EstateRepository extends Repository<EstateEntity> {
